Config Annotations+Gui and Config-Autosync from Server.
This switches the Config over to Config Annotations, whilst at the same time remaining compatible with the existing Config Files. This results in the automatic creating of a Config Gui.

The Config functions In-Game as a master copy for the SyncedConfig class, which holds all the config values and automatically syncs them from Server to Client. In order to do this the FieldSerializer class is responsible for converting the Fields values into a syncable Data Format. Because of the TypeMapping performed in the Process it now becomes possible to use any POJO as a synced field in the SyncedConfig class, which can be mapped to a single valid Config Entry. This PR makes use of this to replace the old BlacklistBlocks class with the new PatternList, which can be accessed directly from the SyncedConfig class.

The PatternList class enables the BG-Block-Blacklist to use Regex. as Entries, so that users can define arbitarily complex Patterns of Blocks to Blacklist (and turn this into a whitelist, with little effort).

In the Process this solves #9...