Goodbye Registry mess, hello clean and simple code
I've made the best solutions that I could think of to not introduce code dupe but also to not over complicate a fairly version specific set of code (basically it doesn't change on each version)

I've also done a lot of work on consistency. I've tried to comment the things that need comments and keep things in a logical order. My only issues at this point is that I think the OurItems builder could be removed.